requirements.txt
praw
moviepy==1.0.3
edge-tts
python-dotenv
requests
pillow
numpy
